+++ src/compas/geometry/brep/brep.py
@@ -1,114 +1,7 @@
 from compas.geometry import Geometry
-from compas.plugins import pluggable
-from compas.plugins import PluginNotInstalledError
 
 
 LINEAR_DEFLECTION = 1e-3
-
-
-@pluggable(category="factories")
-def new_brep(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_boolean_difference(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_boolean_intersection(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_boolean_union(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_box(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_brepfaces(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_cone(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_curves(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_cylinder(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_extrusion(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_iges(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_loft(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_mesh(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_native(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_pipe(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_polygons(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_sphere(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_step(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_surface(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_sweep(*args, **kwargs):
-    raise PluginNotInstalledError()
-
-
-@pluggable(category="factories")
-def from_torus(*args, **kwargs):
-    raise PluginNotInstalledError()
 
 
 class BrepType(object):
